refactor(web): extract the deploy/publish modal into a dumb DeployModal component
Continues Cluster E of the FileViewer.tsx decomposition: the deploy modal's
portal/backdrop/provider-form/Cloudflare-zone-picker/result-cards/social-share
JSX moves into features/file-viewer/components/DeployModal.tsx (props in,
JSX out), wired from FileViewer.tsx with flat props sourced from the
already-landed useWiredDeployFlow hook. The portal target is a portalRoot
prop instead of a bare document read, since the guard forbids document
inside features/**. Structural move only, no logic/markup changes.

Still inline: the share-menu dropdown and its toolbar trigger.
